Fix formatting string_view in UTF-8 build
Don't use string_view data directly, as this doesn't respect its length
and would use the entire rest of the string this view is based on.

Instead, make a copy of just the part corresponding in the view to
ensure that it is NUL-terminated and also use a temporary buffer to hold
it to ensure that it lives long enough.